remove old network chart
