+++ src/features/metrics/components/networks/networks.area-chart.tsx
@@ -1,43 +1,52 @@
 import Card from "@/components/card";
-
-import useServerEventsContext from "@/hooks/useServerEventsContext";
 import formatBytes from "@/features/metrics/utils/format-bytes";
-import AreaChart, { DatasetOptions } from "@/components/area-chart";
+import AreaChart, { useAreaChartState } from "@/components/area-chart.prototype";
+import useServerEventsContext from "@/hooks/useServerEventsContext";
+import { useEffect } from "react";
+import { Options, SeriesOptionsType } from "highcharts";
+
+// TODO: Remove Luxon and ChartJS
+// TODO: Make timestamp work automatically
+// TODO: fix time
 
 const NetworksAreaChart: React.FC = ({}) => {
   const { networks } = useServerEventsContext();
-  const title = "Network Received";
-  const labels = networks?.map((network) => network.data[0].timestamp);
-  const datasets: DatasetOptions[] = networks?.map((network) => ({
-    label: network.id,
-    data: network.data.map((data) => ({
-      x: data.timestamp,
-      y: data.received,
-    })),
-    backgroundColor: "rgba(255,215,120,0.4)",
-    borderColor: "rgba(255,215,120,1)",
-    fill: true,
-    yAxisId: "network-received",
-  }));
-  const callbacks = {
-    label: (context: any) => {
-      const label = context.dataset.label || "";
-      const value = formatBytes(context.parsed.y);
-      return `${label}: ${value}`;
+  const [chartOptions, setChartOptions] = useAreaChartState({
+    title: {
+      text: "Network Received",
     },
-  };
+    yAxis: {
+      labels: {
+        formatter: (x) => formatBytes(x.value as number),
+      },
+    },
+    tooltip: {
+      pointFormatter: function () {
+        return `<span style="color:${this.color}">\u25CF</span> ${this.series.name}: <b>${formatBytes(
+          this.y as number
+        )}</b><br/>`;
+      },
+    },
+  });
 
-  const yAxisTicksCallback = (value: number) => formatBytes(value);
+  useEffect(() => {
+    setChartOptions({
+      series: networks.reduce<SeriesOptionsType[]>((acc, network) => {
+        return [
+          ...acc,
+          {
+            name: `${network.id}`,
+            type: "area",
+            data: network.data.map((net) => [net.timestamp, net.received]),
+          },
+        ];
+      }, []),
+    });
+  }, [networks]);
 
   return (
-    <Card style={{ height: "300px" }}>
-      <AreaChart
-        title={title}
-        labels={labels}
-        datasets={datasets}
-        callbacks={callbacks}
-        yAxisTicksCallback={yAxisTicksCallback}
-      />
+    <Card style={{ height: "450px" }}>
+      <AreaChart options={chartOptions} />
     </Card>
   );
 };
