+++ src/features/metrics/pages/dashboard.page.tsx
@@ -2,7 +2,7 @@ import SystemInfo from "@/features/metrics/components/system-info";
 import MemoryAreaChart from "@/features/metrics/components/memory/memory.area-chart";
 import SwapAreaChart from "@/features/metrics/components/swap/swap.area-chart";
 import GlobalCpuAreaChart from "@/features/metrics/components/global-cpu/global-cpu.area-chart";
-import NetworksAreaChart from "@/features/metrics/components/networks/networks.area-chart.prototype";
+import NetworksAreaChart from "@/features/metrics/components/networks/networks.area-chart";
 import MemoryRadialChart from "@/features/metrics/components/memory/memory.radial-chart";
 import SwapRadialChart from "@/features/metrics/components/swap/swap.radial-chart";
 import CpusBarChart from "@/features/metrics/components/cpus/cpus.bar-chart";
